+++ shiki_organizer/cli/interval.py
@@ -2,10 +2,11 @@ import datetime as dt
 import uuid
 
 import click
-from colorama import Fore, Style
 
 from shiki_organizer.model import Interval, Task
-from shiki_organizer.formatting import interval_to_str
+from shiki_organizer.formatting import duration_to_str, interval_to_str
+import pendulum
+from colorama import Fore, Style
 
 
 @click.group()
@@ -17,9 +18,37 @@ def cli():
 @click.option(
     "-u", "--uuid", is_flag=True, default=False, help="Show uuid of interval."
 )
-def ls(uuid):
-    for interval in Interval.select():
+@click.option(
+    "-p",
+    "--period",
+    type=click.Choice(["all", "today", "week", "month", "year"]),
+    default="all",
+    help="Show data only for this period of time.",
+)
+def ls(uuid, period):
+    start = pendulum.now()
+    match period:
+        case "today":
+            start = start.start_of("day")
+        case "week":
+            start = start.start_of("week")
+        case "month":
+            start = start.start_of("month")
+        case "year":
+            start = start.start_of("year")
+        case _:
+            start = dt.datetime.min
+    if start != dt.datetime.min:
+        start_string = start.to_datetime_string()
+        start = dt.datetime.fromisoformat(start_string)
+    intervals = Interval.select().where(Interval.start >= start)
+    total_duration = 0
+    for interval in intervals:
+        total_duration += interval.duration
         print(interval_to_str(interval, uuid))
+    print(
+        f"{Fore.MAGENTA}Total duration:{Style.RESET_ALL}{duration_to_str(total_duration)}"
+    )
 
 
 @click.command()